#1: Write to the console the sum of the areas.

+++ 01coreSikidomokTerulete/Circle.cs
@@ -11,9 +11,9 @@ namespace _01coreSikidomokTerulete
             this.radius = radius;
         }
 
-        public double Area()
+        public int Area()
         {
-            return Math.PI * radius * radius;
+            return (int) Math.PI * radius * radius;
         }
     }
 }
